fix : SyncJob 조건 기반 count 쿼리(QueryDSL) 기능 추가
- SyncJob 데이터의 조건 기반 전체 개수(totalElements)를 조회하는 기능 추가
- 커스텀 레포지토리 인터페이스(SyncJobRepositoryCountCustom) 및 구현체(SyncJobRepositoryCountCustomImpl) 분리 작성
- QueryDSL을 사용하여 직접 count 쿼리 작성
- CursorWhere.booleanBuilder를 재사용하여 조건 빌더 통합 적용

+++ src/main/java/com/findex/demo/autoSyncConfig/repository/AutoSyncConfigRepositoryImpl.java
@@ -57,7 +57,7 @@ public class AutoSyncConfigRepositoryImpl implements
                 null,
                 nextIdAfter,
                 size,
-                result.size(),
+                (long) result.size(),
                 hasNext
         );
     }
